+++ packages/klab-pytest-toolkit-embedded/src/klab_pytest_toolkit_embedded/debug_probes/__init__.py
@@ -1,4 +1,6 @@
 from klab_pytest_toolkit_embedded.debug_probes.interface import DebugProbe
 from klab_pytest_toolkit_embedded.debug_probes.esp import EspTool
+from klab_pytest_toolkit_embedded.debug_probes.openocd import OpenOcdProbe
+from klab_pytest_toolkit_embedded.debug_probes.probers import ProbeRsProbe
 
-__all__ = ["DebugProbe", "EspTool"]
+__all__ = ["DebugProbe", "EspTool", "OpenOcdProbe", "ProbeRsProbe"]