+++ packages/klab-pytest-toolkit-embedded/src/klab_pytest_toolkit_embedded/gpio_controllers/__init__.py
@@ -1,0 +1,4 @@
+from klab_pytest_toolkit_embedded.gpio_controllers.ftdi import FtdiGpioController
+from klab_pytest_toolkit_embedded.gpio_controllers.interface import GpioController
+
+__all__ = ["FtdiGpioController", "GpioController"]